+++ src/packages/core/collection/conditions/writable/constants.ts
