+++ src/packages/core/collection/conditions/writable/manifests.ts
