+++ src/packages/core/collection/manifests.ts
@@ -1,6 +1,6 @@
 import { manifest as collectionAliasCondition } from './collection-alias.manifest.js';
 import { manifest as collectionBulkActionPermissionCondition } from './collection-bulk-action-permission.manifest.js';
-import { manifests as conditionManifests } from './conditions/manifests.js';
+import { manifests as conditionManifests } from './conditions/writable/manifests.js';
 import type { ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
 
 export const manifests: Array<ManifestTypes> = [
